backend/requirements.txt
fastapi
docker
uvicorn
sqlalchemy
psycopg2-binary
bcrypt==4.0.1
python-dotenv
python-owasp-zap-v2.4
slack_sdk
selenium
jinja2
websockets
pika
PyYAML
xhtml2pdf
reportlab
nltk
beautifulsoup4
fuzzywuzzy
passlib
python-jose
pycryptodome
PyJWT
python-multipart
mechanize
webdriver-manager
matplotlib
email-validator
pydantic[email]